requirements.txt
flask==3.0.0
gunicorn==21.2.0
openai-whisper
numpy
